requirements.txt
pandapower>=2.13
numpy>=1.24
scipy>=1.10
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3
torch>=2.0
gymnasium>=0.29
tqdm>=4.65
PyYAML>=6.0
